requirements.txt
requests
sentence_transformers
qdrant_client
python-dotenv